requirements.txt
python-fasthtml
pyyaml
pytz
feedgen
markdown
rich

# Specials
nb2fasthtml
fh-frankenui # To be made optional